Fold BlobExtensions.cs into Blob.cs
