+++ LibGit2Sharp/Blob.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Text;
 using LibGit2Sharp.Core;
 
 namespace LibGit2Sharp
@@ -55,7 +56,64 @@ namespace LibGit2Sharp
         public virtual Stream GetContentStream(FilteringOptions filteringOptions)
         {
             Ensure.ArgumentNotNull(filteringOptions, "filteringOptions");
+
             return Proxy.git_blob_filtered_content_stream(repo.Handle, Id, filteringOptions.HintPath, false);
+        }
+
+        /// <summary>
+        /// Gets the blob content, decoded with UTF8 encoding if the encoding cannot be detected from the byte order mark
+        /// </summary>
+        /// <returns>Blob content as text.</returns>
+        public virtual string GetContentText()
+        {
+            return ReadToEnd(GetContentStream(), null);
+        }
+
+        /// <summary>
+        /// Gets the blob content decoded with the specified encoding,
+        /// or according to byte order marks, or the specified encoding as a fallback
+        /// </summary>
+        /// <param name="encoding">The encoding of the text to use, if it cannot be detected</param>
+        /// <returns>Blob content as text.</returns>
+        public virtual string GetContentText(Encoding encoding)
+        {
+            Ensure.ArgumentNotNull(encoding, "encoding");
+
+            return ReadToEnd(GetContentStream(), encoding);
+        }
+
+        /// <summary>
+        /// Gets the blob content, decoded with UTF8 encoding if the encoding cannot be detected
+        /// </summary>
+        /// <param name="filteringOptions">Parameter controlling content filtering behavior</param>
+        /// <returns>Blob content as text.</returns>
+        public virtual string GetContentText(FilteringOptions filteringOptions)
+        {
+            return GetContentText(filteringOptions, null);
+        }
+
+        /// <summary>
+        /// Gets the blob content as it would be checked out to the
+        /// working directory, decoded with the specified encoding,
+        /// or according to byte order marks, with UTF8 as fallback,
+        /// if <paramref name="encoding"/> is null.
+        /// </summary>
+        /// <param name="filteringOptions">Parameter controlling content filtering behavior</param>
+        /// <param name="encoding">The encoding of the text. (default: detected or UTF8)</param>
+        /// <returns>Blob content as text.</returns>
+        public virtual string GetContentText(FilteringOptions filteringOptions, Encoding encoding)
+        {
+            Ensure.ArgumentNotNull(filteringOptions, "filteringOptions");
+
+            return ReadToEnd(GetContentStream(filteringOptions), encoding);
+        }
+
+        private static string ReadToEnd(Stream stream, Encoding encoding)
+        {
+            using (var reader = new StreamReader(stream, encoding ?? LaxUtf8Marshaler.Encoding, encoding == null))
+            {
+                return reader.ReadToEnd();
+            }
         }
     }
 }
